Rename Sticky Activities sample
