+++ src/WorkerSpecificTaskQueues/FileProcessingWorkflow.workflow.cs
@@ -1,6 +1,6 @@
 using Temporalio.Workflows;
 
-namespace TemporalioSamples.ActivityStickyQueues;
+namespace TemporalioSamples.WorkerSpecificTaskQueues;
 
 [Workflow]
 public class FileProcessingWorkflow
@@ -13,7 +13,7 @@ public class FileProcessingWorkflow
             new() { StartToCloseTimeout = TimeSpan.FromMinutes(1) });
 
         var downloadPath = await Workflow.ExecuteActivityAsync(
-            () => StickyActivities.DownloadFileToWorkerFileSystemAsync("https://temporal.io"),
+            () => WorkerSpecificActivities.DownloadFileToWorkerFileSystemAsync("https://temporal.io"),
             new()
             {
                 TaskQueue = uniqueWorkerTaskQueue,
@@ -26,7 +26,7 @@ public class FileProcessingWorkflow
             });
 
         await Workflow.ExecuteActivityAsync(
-            () => StickyActivities.WorkOnFileInWorkerFileSystemAsync(downloadPath),
+            () => WorkerSpecificActivities.WorkOnFileInWorkerFileSystemAsync(downloadPath),
             new()
             {
                 TaskQueue = uniqueWorkerTaskQueue,
@@ -35,7 +35,7 @@ public class FileProcessingWorkflow
             });
 
         await Workflow.ExecuteActivityAsync(
-            () => StickyActivities.CleanupFileFromWorkerFileSystemAsync(downloadPath),
+            () => WorkerSpecificActivities.CleanupFileFromWorkerFileSystemAsync(downloadPath),
             new()
             {
                 TaskQueue = uniqueWorkerTaskQueue,
